requirements.txt
Django>=4.2,<5.0
django-filter>=23.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
gunicorn>=21.2.0
drf-spectacular